refactor: ContentHeader 컴포넌트 분리 #448
